+++ src/main/java/http/response/HttpResponse.java
@@ -4,40 +4,40 @@ import http.HttpHeaders;
 
 public class HttpResponse {
 
-	private ResponseLine responseLine;
-	private ContentType contentType;
-	private HttpHeaders httpHeaders;
-	private byte[] body;
+    private StatusLine statusLine;
+    private ContentType contentType;
+    private HttpHeaders httpHeaders;
+    private byte[] body;
 
-	public ResponseLine getResponseLine() {
-		return responseLine;
-	}
+    public StatusLine getResponseLine() {
+        return statusLine;
+    }
 
-	public void setResponseLine(ResponseLine responseLine) {
-		this.responseLine = responseLine;
-	}
+    public void setResponseLine(StatusLine statusLine) {
+        this.statusLine = statusLine;
+    }
 
-	public ContentType getContentType() {
-		return contentType;
-	}
+    public ContentType getContentType() {
+        return contentType;
+    }
 
-	public void setContentType(ContentType contentType) {
-		this.contentType = contentType;
-	}
+    public void setContentType(ContentType contentType) {
+        this.contentType = contentType;
+    }
 
-	public HttpHeaders getHttpHeaders() {
-		return httpHeaders;
-	}
+    public HttpHeaders getHttpHeaders() {
+        return httpHeaders;
+    }
 
-	public void setHttpHeaders(HttpHeaders httpHeaders) {
-		this.httpHeaders = httpHeaders;
-	}
+    public void setHttpHeaders(HttpHeaders httpHeaders) {
+        this.httpHeaders = httpHeaders;
+    }
 
-	public byte[] getBody() {
-		return body;
-	}
+    public byte[] getBody() {
+        return body;
+    }
 
-	public void setBody(byte[] body) {
-		this.body = body;
-	}
+    public void setBody(byte[] body) {
+        this.body = body;
+    }
 }
